[Partisia] Move sample data utilities out of production source set - #113
