+++ document-manager/src/test/java/eu/europa/ec/eudi/wallet/document/sample/SampleDocumentManager.kt
@@ -139,4 +139,3 @@ interface SampleDocumentManager : DocumentManager {
         }
     }
 }
-
